pyproject.toml
[project]
name = "gentle-man-tg-bot"
version = "1.0.0"
description = "Intelligent Telegram bot with Gemini Flash 2.0, FAISS vector search, and SQLite"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "click>=8.1.8",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "pyyaml>=6.0.0",
    "sqlalchemy>=2.0.0",
    "greenlet>=3.0.0",
    "openai>=1.75.0",
    "faiss-cpu>=1.9.0",
    "numpy>=2.2.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "coverage>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 150
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 150

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true